Move search redirects to own module. Add blog redirect.
