+++ tests/store/test_views.py
@@ -34,54 +34,6 @@ class StoreViews(TestCase):
         response = self.client.get(url_for("jaasstore.search", q="k8s"))
         self.assertEqual(response.status_code, 200)
 
-    @patch("webapp.store.models.search_entities")
-    def test_search_redirect(self, mock_search_entities):
-        mock_search_entities.return_value = []
-        response = self.client.get("/q/k8s")
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(
-            response.location,
-            "http://localhost{}".format(url_for("jaasstore.search", q="k8s")),
-        )
-
-    @patch("webapp.store.models.search_entities")
-    def test_search_redirect_multi_word(self, mock_search_entities):
-        mock_search_entities.return_value = []
-        response = self.client.get("/q/k8s/demo")
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(
-            response.location,
-            "http://localhost{}".format(
-                url_for("jaasstore.search", q="k8s demo")
-            ),
-        )
-
-    @patch("webapp.store.models.search_entities")
-    def test_search_redirect_query_params(self, mock_search_entities):
-        mock_search_entities.return_value = []
-        response = self.client.get("/q/k8s?sort=name&series=xenial")
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(
-            response.location,
-            "http://localhost{}".format(
-                url_for(
-                    "jaasstore.search", q="k8s", sort="name", series="xenial"
-                )
-            ),
-        )
-
-    @patch("webapp.store.models.search_entities")
-    def test_search_redirect_just_query_params(self, mock_search_entities):
-        mock_search_entities.return_value = []
-        response = self.client.get("/q?tags=proxy")
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(
-            response.location,
-            "http://localhost{}".format(
-                url_for("jaasstore.search", tags="proxy")
-            ),
-        )
-
     @patch("webapp.store.models.get_user_entities")
     def test_user_details(self, mock_get_user_entities):
         mock_get_user_entities.return_value = {
