+++ webapp/store/views.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, abort, redirect, request, render_template
+from flask import Blueprint, abort, request, render_template
 from webapp.store import models
 
 from jujubundlelib import references
@@ -52,22 +52,6 @@ def search():
             "query": query,
         },
     )
-
-
-@jaasstore.route("/q/")
-@jaasstore.route("/q/<path:path>")
-def search_redirect(path=None):
-    """
-    Handle redirects from jujucharms.com search URLS to the jaas.ai format.
-    e.g. /q/k8s/demo?sort=-name&series=xenial will redirect to
-    /search?q=k8s+demo&sort=-name&series=xenial
-    """
-    query_string = []
-    if path:
-        query_string.append("q={}".format(path.replace("/", "+")))
-    if request.query_string:
-        query_string.append(str(request.query_string, "utf-8"))
-    return redirect("/search?{}".format("&".join(query_string)), code=302)
 
 
 @jaasstore.route("/u/<username>")
